feat:SGED-87: Add deleteMissionMiddleware, use http delete method to delete mission

+++ src/controllers/missionController.ts
@@ -148,7 +148,7 @@ export const updateMission = async (req: Request, res: Response): Promise<void> 
     let previousMission: Mission | null = null;
 
     try {
-        previousMission = await missionRepository.getMission(mission.id)
+        previousMission = await missionRepository.getMission(mission.id, missionTest)
         const previousMissionManagerIds = await missionRepository.getMissionManagers(mission.id)
             .then(managers => managers.map(m => m.userId));
         const previousMissionManagerIdsSet = new Set(previousMissionManagerIds);
@@ -207,11 +207,8 @@ export const updateMission = async (req: Request, res: Response): Promise<void> 
 }
 
 export const deleteMission = async (req: Request, res: Response): Promise<void> => {
+    const missionId = req.params.missionId;
     const missionTest = req.claims?.tester ?? false;
-    const {
-        MISSION_ID: missionId,
-        MISSION_IMAGE_FILE_NAME: imageFileName
-    } = req.body;
 
     if(!missionId) {
         res.status(400).json(badRequestErrorResponse);
@@ -219,13 +216,20 @@ export const deleteMission = async (req: Request, res: Response): Promise<void> 
     }
 
     try {
+        const mission = await missionRepository.getMission(missionId, missionTest);
+
+        if (!mission) {
+            res.status(204);
+            return;
+        }
+
         await missionRepository.deleteMission(missionId, missionTest);
         const serverResponse: ServerResponse = { message: 'Mission has been deleted successfully' };
         res.status(200).json(serverResponse);
 
-        if (imageFileName) {
-            await imageRepository.deleteImage(getImagePath(imageFileName))
-                .catch(error => e(new Error(`Error deleting mission image ${imageFileName} : ${error.message}`)));
+        if (mission.imageFileName) {
+            await imageRepository.deleteImage(getImagePath(mission.imageFileName))
+                .catch(error => e(new Error(`Error deleting mission image ${mission.imageFileName} : ${error.message}`)));
         }
     } catch (error: any) {
         e(new Error(`Error deleting mission ${missionId} : ${error.message}`));
